Refactor _annotate_observable_at_node and _get_ordered_leaves
